extract consul's version normalizer into new input package to be shared across resources
