+++ internal/consul/version_test.go
@@ -117,31 +117,6 @@ func Test_IsValidVersion(t *testing.T) {
 	}
 }
 
-func Test_NormalizeVersion(t *testing.T) {
-	tcs := map[string]struct {
-		expected string
-		input    string
-	}{
-		"with a prefixed v": {
-			input:    "v1.9.0",
-			expected: "v1.9.0",
-		},
-		"without a prefixed v": {
-			input:    "1.9.0",
-			expected: "v1.9.0",
-		},
-	}
-
-	for n, tc := range tcs {
-		t.Run(n, func(t *testing.T) {
-			r := require.New(t)
-
-			result := NormalizeVersion(tc.input)
-			r.Equal(tc.expected, result)
-		})
-	}
-}
-
 func Test_VersionsToString(t *testing.T) {
 	tcs := map[string]struct {
 		expected string
