+++ internal/input/input.go
@@ -1,0 +1,10 @@
+package input
+
+import (
+	"strings"
+)
+
+// NormalizeVersion ensures the version starts with a 'v'
+func NormalizeVersion(version string) string {
+	return "v" + strings.TrimPrefix(version, "v")
+}
